test(e2e): split rewards, apply cleancode optimizations
- Move buildClaimBorrowRewardsTransaction to rewards.test.ts for consistency
- Fix vaults assertions: use toBe('bigint') and toBeGreaterThanOrEqual
- Replace early returns with explicit expect in oracle tests
- Extract TEST_AMOUNT_USDB constant in saving.test.ts

Made-with: Cursor
